restructure for distribution
