+++ page-uploader/app/assets/css/styles.css
