--- page-uploader/keys.js
@@ -1,20 +1,0 @@
-let credentials;
-
-function setKeys(keys) {
-	credentials = keys;
-}
-
-function getSecret() {
-	return credentials.secret || null;
-}
-
-function getKey() {
-	return credentials.key || null;
-}
-
-
-module.exports = {
-	set: setKeys,
-	secret: getSecret,
-	key: getKey
-}
